cldr-emoji-annotation: update to 33.0.0.1.
pkgsrc changes:

- use gmake to handle upstream configure.ac changes

Upstream announcement:
 http://cldr.unicode.org/index/downloads/cldr-33

Unicode CLDR 33 provides an update to the key building blocks for
software supporting the world's languages. This data is used by all
major software systems for their software internationalization and
localization, adapting software to the conventions of different
languages for such common software tasks.

This release had a limited submission phase. The focus was on
improvements to emoji keywords and to the Odia and Assamese locales,
addition of typographic names data, and improvements to the structure
for specifying keyboard layouts.