devel/go-gocode: Import version 20170907
Gocode is a helper tool which is intended to be integrated with your
source code editor, like vim, neovim and emacs. It provides several
advanced capabilities, which currently includes context-sensitive
autocompletion.  It is called daemon, because it uses client/server
architecture for caching purposes. In particular, it makes
autocompletions very fast. Typical autocompletion time with warm cache
is 30ms, which is barely noticeable.