x11/caribou: import caribou-0.4.21
Caribou is a text entry and UI navigation application being developed as an
alternative to the Gnome On-screen Keyboard. The overarching goal for Caribou is
to create a usable solution for people whose primary way of accessing a computer
is a switch device.

The initial goal is to make an in-place on-screen keyboard suitable for people
who can use a mouse but not a hardware keyboard. This on-screen keyboard may
also be useful for touch screen or tablet users.

The GNOME Shell onscreen keyboard uses libcaribou.